Added fixed random_seed and better checks to test_is_invertible_module, also added more tests
* Added better tests for is_invertible_module
* Moved most of the test models to test_models.py
* Removed TODO for injecting gradients, since this should become a new issue
